minor cleanups and moved texture changing helper methods to stateful_object
