+++ igibson/objects/stateful_object.py
@@ -1,14 +1,24 @@
-import sys
 import logging
-import numpy as np
-from collections import OrderedDict
+import os
+import sys
+from collections import OrderedDict, defaultdict
 
+import numpy as np
 import omni
-from igibson.object_states.factory import get_state_name, get_default_states, get_steam_states, get_states_for_ability, get_object_state_instance
-from igibson.object_states.object_state_base import CachingEnabledObjectState, REGISTERED_OBJECT_STATES
+from omni.isaac.core.utils.prims import get_prim_at_path
+from omni.usd import get_shader_from_material
+from pxr.Sdf import ValueTypeNames as VT
+
+from igibson.object_states.factory import (
+    get_default_states,
+    get_object_state_instance,
+    get_state_name,
+    get_states_for_ability,
+    get_steam_states,
+)
+from igibson.object_states.object_state_base import REGISTERED_OBJECT_STATES, CachingEnabledObjectState
 from igibson.objects.object_base import BaseObject
 from igibson.renderer_settings.renderer_settings import RendererSettings
-from pxr.Sdf import ValueTypeNames as VT
 
 # Optionally import bddl for object taxonomy.
 try:
@@ -63,7 +73,8 @@ class StatefulObject(BaseObject):
         """
         # Values that will be filled later
         self._states = None
-        self._has_steam_prims = False
+        self._texture_states = []
+        self._extra_prims_flags = defaultdict(bool)
         self._emitter = None
 
         # Load abilities from taxonomy if needed & possible
@@ -106,7 +117,7 @@ class StatefulObject(BaseObject):
             state.initialize(self._simulator)
             
         # Create steam prims if this object has states with steam effects.
-        if self._has_steam_prims:
+        if self._extra_prims_flags["has_steam_prims"]:
             self._create_steam_prims()
 
     def initialize_states(self):
@@ -168,7 +179,7 @@ class StatefulObject(BaseObject):
         self.initialize_states()
         for state_type, params in reversed(state_types_and_params):
             if state_type in get_steam_states():
-                self._has_steam_prims = True
+                self._extra_prims_flags["has_steam_prims"] = True
             self._states[state_type] = get_object_state_instance(state_type, self, params)
 
     def _create_steam_prims(self):
@@ -226,6 +237,76 @@ class StatefulObject(BaseObject):
             value (bool): Value to set
         """
         self._emitter.CreateAttribute("enabled", VT.Bool, False).Set(value)
+
+    def get_textures(self):
+        """Gets prim's texture files.
+
+        Returns:
+            list of (str): List of texture file paths
+        """
+        textures = []
+        looks_prim_path = f"{str(self._prim_path)}/Looks"
+        looks_prim = get_prim_at_path(looks_prim_path)
+        if not looks_prim:
+            return
+        for subprim in looks_prim.GetChildren():
+            if subprim.GetPrimTypeInfo().GetTypeName() != "Material":
+                continue
+            shader = get_shader_from_material(subprim)
+            texture_path = shader.GetInput("diffuse_texture").Get()
+            if texture_path:
+                textures.append(texture_path.path)
+        return textures
+
+    def update_textures_for_state(self, state, value):
+        """Update prim's textures to represent @state.
+
+        Args:
+            state (BaseObjectState): State to represent
+            value (bool): Whether or not changing to the state
+        """
+        # Determine the state to set.
+        TEXTURE_CHANGE_PRIORITY = {
+            "Frozen": 4,
+            "Burnt": 3,
+            "Cooked": 2,
+            "Soaked": 1,
+            "ToggledOn": 0,
+        }
+        if value and state.__name__ not in self._texture_states:
+            self._texture_states.append(state.__name__)
+            self._texture_states.sort(key=lambda s: TEXTURE_CHANGE_PRIORITY[s])
+        if not value and state.__name__ in self._texture_states:
+            self._texture_states.remove(state.__name__)
+        
+        state_to_set = self._texture_states[-1] if self._texture_states else None
+
+        # Find the material prims to update.
+        looks_prim_path = f"{str(self._prim_path)}/Looks"
+        looks_prim = get_prim_at_path(looks_prim_path)
+        if not looks_prim:
+            return
+        for subprim in looks_prim.GetChildren():
+            if subprim.GetPrimTypeInfo().GetTypeName() != "Material":
+                continue
+            shader = get_shader_from_material(subprim)
+            texture_path = shader.GetInput("diffuse_texture").Get()
+            if texture_path is None:
+                continue
+            # Get updated texture file path for state.
+            texture_path_split = texture_path.path.split("/")
+            filedir, filename = "/".join(texture_path_split[:-1]), texture_path_split[-1]
+            assert filename[-4:] == ".png", f"Texture file {filename} does not end with .png"
+            filename_split = filename[:-4].split("_")
+            # Check both file names for backward compatibility.
+            if len(filename_split) > 0 and filename_split[-1] not in ("DIFFUSE", "albedo"):
+                filename_split.pop()
+            target_texture_path = f"{filedir}/{'_'.join(filename_split)}"
+            target_texture_path += f"_{state_to_set}.png" if state_to_set else ".png"
+            if not os.path.exists(target_texture_path):
+                print(f"Warning: get texture path failed because {target_texture_path} does not exist")
+                continue
+            shader.GetInput("diffuse_texture").Set(target_texture_path)
 
     def _dump_state(self):
         # Grab state from super class
